Move union_coercion to type_coercion.rs
